[REFACTOR/#27] 폴더링 변경
