+++ app/src/main/java/com/growthook/aos/presentation/cavedetail/CaveDetailSelectMenuBottomSheet.kt
@@ -10,7 +10,7 @@ import android.widget.Toast
 import androidx.fragment.app.activityViewModels
 import com.growthook.aos.R
 import com.growthook.aos.databinding.FragmentSelectMenuBottomsheetBinding
-import com.growthook.aos.presentation.insight.noactionplan.InsightMenuBottomsheet
+import com.growthook.aos.presentation.insight.noactionplan.see.InsightMenuBottomsheet
 import com.growthook.aos.util.base.BaseAlertDialog
 import com.growthook.aos.util.base.BaseBottomSheetFragment
 import dagger.hilt.android.AndroidEntryPoint
