+++ app/src/main/java/com/growthook/aos/presentation/home/HomeFragment.kt
@@ -18,8 +18,8 @@ import com.growthook.aos.domain.entity.Cave
 import com.growthook.aos.domain.entity.Insight
 import com.growthook.aos.presentation.MainActivity
 import com.growthook.aos.presentation.cavedetail.CaveDetailActivity
-import com.growthook.aos.presentation.insight.noactionplan.InsightMenuBottomsheet
-import com.growthook.aos.presentation.insight.noactionplan.NoActionplanInsightActivity
+import com.growthook.aos.presentation.insight.noactionplan.see.InsightMenuBottomsheet
+import com.growthook.aos.presentation.insight.noactionplan.see.NoActionplanInsightActivity
 import com.growthook.aos.util.EmptyDataObserver
 import com.growthook.aos.util.base.BaseAlertDialog
 import com.growthook.aos.util.base.BaseFragment
