+++ app/src/main/res/layout/activity_seed_modify.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.insight.noactionplan.SeedModifyActivity">
+    tools:context=".presentation.insight.noactionplan.see.SeedModifyActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
