+++ app/src/main/res/layout/fragment_insight_cave_bottomsheet.xml
@@ -6,7 +6,7 @@
     android:layout_height="wrap_content"
     android:background="@color/Gray400"
     android:paddingBottom="18dp"
-    tools:context=".presentation.insight.noactionplan.InsightCaveBottomsheet">
+    tools:context=".presentation.insight.noactionplan.see.InsightCaveBottomsheet">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_insight_cave_main"
